+++ src/main/java/com/zwz/jxone/controller/ModelStrategyController.java
@@ -19,7 +19,7 @@ public class ModelStrategyController {
     @Autowired
     private ModelStrategyService modelStrategyService;
     @RequestMapping(method = RequestMethod.GET)
-    public ModelStrategyPO selectById(@PathVariable Long id){
+    public ModelStrategyPO selectById(@PathVariable Integer id){
         ModelStrategyPO modelStrategyPO=modelStrategyService.selectById(id);
         return modelStrategyPO;
     }
