+++ src/main/java/com/zwz/jxone/dao/mapper/ModelMapper.java
@@ -11,9 +11,9 @@ import org.springframework.stereotype.Repository;
 public interface ModelMapper {
     Integer insertMode();
 
-    ModelPO selectById(Long id);
+    ModelPO selectById(Integer id);
     ModelPO updateById(ModelPO modelPO);
-    ModelPO deleteById(Long id);
+    ModelPO deleteById(Integer id);
     ModelPO insert(ModelPO modelPO);
 
 }
